# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniSync",
    "version": "see VERSION"
  },
  "paths": {
...
    "/remotes": {
      "get": {
        "summary": "List Remotes",
        "description": "Return the list of configured rclone remotes.",
        "operationId": "list_remotes_remotes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RemoteResponse"
                  },
                  "type": "array",
                  "title": "Response List Remotes Remotes Get"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/remotes/import/preview": {
      "post": {
        "summary": "Preview Import",
        "description": "List the remotes of an uploaded rclone.conf: which can be imported,\nwhich clash with an existing remote of the same name, which are refused.",
        "operationId": "preview_import_remotes_import_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreviewResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/remotes/import": {
      "post": {
        "summary": "Import Remotes",
        "description": "Add the selected remotes of an uploaded rclone.conf, all or none.\n\nEach selection names a section of the file and the name to add it under\n(to resolve a clash with an existing remote). Values are copied verbatim;\nreferences between imported remotes (a crypt remote's ``remote``)\nfollow a rename. A clash is 409, a refused section or name 422.",
        "operationId": "import_remotes_remotes_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRemotesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportRemotesResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/remotes/{name}/config": {
      "get": {
        "summary": "Get Remote Config",
        "description": "The remote's settings for the edit form. Secrets (passwords, keys,\nclient secrets) only say whether they are set; tokens are not included.",
        "operationId": "get_remote_config_remotes__name__config_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoteConfigResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/remotes/{name}": {
      "put": {
        "summary": "Update Remote",
        "description": "Change a remote's settings without deleting it.\n\nValidated like a new remote (only the provider's fields, see\nvalidate_remote_params). Fields left out stay as they are; a non-secret\nfield set to \"\" is removed; a secret set to \"\" is kept, another value\nreplaces it (obscured where rclone expects that); ``clear`` removes\nsecrets. OAuth remotes change through a reconnect instead.",
        "operationId": "update_remote_remotes__name__put",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRemoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Update Remote Remotes  Name  Put"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Remote",
        "description": "Delete an rclone remote. Returns 409 if dependencies exist unless force=true.",
        "operationId": "delete_remote_remotes__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Remote Remotes  Name  Delete"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/remotes/{name}/about": {
      "get": {
        "summary": "Get Remote About",
        "description": "Get storage info for a remote.",
        "operationId": "get_remote_about_remotes__name__about_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoteStorageInfoResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/remotes/{name}/test": {
      "post": {
        "summary": "Test Remote",
        "description": "Test connectivity to a remote by running 'rclone lsd'.",
        "operationId": "test_remote_remotes__name__test_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoteTestResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "What went wrong, for people. Show this text."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable code (snake_case) to branch on."
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Extra data for some codes, e.g. errors, invalid_paths, names, replacement, retry_after."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "The request's id (also the X-Request-ID header); the server log lines of the request carry it."
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "description": "The body of every error answer (4xx and 5xx); docs/api-errors.md lists the codes."
      },
...
      "RemoteResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "last_verified": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "editable": {
            "type": "boolean",
            "title": "Editable",
            "default": false
          },
          "reconnectable": {
            "type": "boolean",
            "title": "Reconnectable",
            "default": false
          },
          "auth_error": {
            "type": "boolean",
            "title": "Auth Error",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "title": "RemoteResponse"
      },
...
    }
  }
}